+++ sdk/core/core-http/test/policies/tracingPolicyTests.ts
@@ -10,9 +10,6 @@ import {
   RequestPolicyOptions
 } from "../../src/coreHttp";
 import {
-  setTracer,
-  NoOpTracer,
-  SpanOptions,
   SpanContext,
   TraceFlags,
   TraceState,
@@ -20,12 +17,11 @@ import {
   context,
   SpanStatusCode,
   SpanStatus,
-  Span,
   SpanAttributes,
-  SpanAttributeValue,
-  Tracer
+  SpanAttributeValue
 } from "@azure/core-tracing";
 import { tracingPolicy } from "../../src/policies/tracingPolicy";
+import { TracerProvider, Tracer, Span, SpanOptions, trace } from "@opentelemetry/api";
 
 class MockSpan implements Span {
   private _endCalled = false;
@@ -128,6 +124,10 @@ class MockTracer implements Tracer {
     private state = ""
   ) {}
 
+  startActiveSpan(): never {
+    throw new Error("Method not implemented.");
+  }
+
   getStartedSpans(): MockSpan[] {
     return this.spans;
   }
@@ -144,10 +144,31 @@ class MockTracer implements Tracer {
   }
 }
 
+class MockTracerProvider implements TracerProvider {
+  private mockTracer: Tracer = new MockTracer();
+
+  setTracer(tracer: Tracer) {
+    this.mockTracer = tracer;
+  }
+
+  getTracer(): Tracer {
+    return this.mockTracer;
+  }
+
+  register() {
+    trace.setGlobalTracerProvider(this);
+  }
+
+  disable() {
+    trace.disable();
+  }
+}
+
 const ROOT_SPAN = new MockSpan("root", "root", TraceFlags.SAMPLED, "");
 
 describe("tracingPolicy", function() {
   const TRACE_VERSION = "00";
+  const mockTracerProvider = new MockTracerProvider();
 
   const mockPolicy: RequestPolicy = {
     sendRequest(request: WebResource): Promise<HttpOperationResponse> {
@@ -159,9 +180,16 @@ describe("tracingPolicy", function() {
     }
   };
 
+  beforeEach(() => {
+    mockTracerProvider.register();
+  });
+
+  afterEach(() => {
+    mockTracerProvider.disable();
+  });
+
   it("will not create a span if spanOptions are missing", async () => {
     const mockTracer = new MockTracer();
-    setTracer(mockTracer);
     const request = new WebResource();
     const policy = tracingPolicy().create(mockPolicy, new RequestPolicyOptions());
     await policy.sendRequest(request);
@@ -173,7 +201,8 @@ describe("tracingPolicy", function() {
     const mockTraceId = "11111111111111111111111111111111";
     const mockSpanId = "2222222222222222";
     const mockTracer = new MockTracer(mockTraceId, mockSpanId, TraceFlags.SAMPLED);
-    setTracer(mockTracer);
+    mockTracerProvider.setTracer(mockTracer);
+
     const request = new WebResource();
     request.tracingContext = setSpan(context.active(), ROOT_SPAN);
 
@@ -199,7 +228,8 @@ describe("tracingPolicy", function() {
     const mockSpanId = "2222222222222222";
     // leave out the TraceOptions
     const mockTracer = new MockTracer(mockTraceId, mockSpanId);
-    setTracer(mockTracer);
+    mockTracerProvider.setTracer(mockTracer);
+
     const request = new WebResource();
     request.tracingContext = setSpan(context.active(), ROOT_SPAN);
 
@@ -227,7 +257,7 @@ describe("tracingPolicy", function() {
     const mockSpanId = "2222222222222222";
     const mockTraceState = "foo=bar";
     const mockTracer = new MockTracer(mockTraceId, mockSpanId, TraceFlags.SAMPLED, mockTraceState);
-    setTracer(mockTracer);
+    mockTracerProvider.setTracer(mockTracer);
     const request = new WebResource();
     request.tracingContext = setSpan(context.active(), ROOT_SPAN);
 
@@ -255,7 +285,7 @@ describe("tracingPolicy", function() {
     const mockSpanId = "2222222222222222";
     const mockTraceState = "foo=bar";
     const mockTracer = new MockTracer(mockTraceId, mockSpanId, TraceFlags.SAMPLED, mockTraceState);
-    setTracer(mockTracer);
+    mockTracerProvider.setTracer(mockTracer);
     const request = new WebResource();
     request.tracingContext = setSpan(context.active(), ROOT_SPAN);
 
@@ -298,8 +328,23 @@ describe("tracingPolicy", function() {
   });
 
   it("will not set headers if span is a NoOpSpan", async () => {
-    setTracer(new NoOpTracer());
+    mockTracerProvider.disable();
     const request = new WebResource();
+
+    const policy = tracingPolicy().create(mockPolicy, new RequestPolicyOptions());
+    await policy.sendRequest(request);
+
+    assert.notExists(request.headers.get("traceparent"));
+    assert.notExists(request.headers.get("tracestate"));
+  });
+
+  it("will not set headers if context is invalid", async () => {
+    // This will create a tracer that produces invalid trace-id and span-id
+    const mockTracer = new MockTracer("invalid", "00", TraceFlags.SAMPLED, "foo=bar");
+    mockTracerProvider.setTracer(mockTracer);
+
+    const request = new WebResource();
+    request.tracingContext = setSpan(context.active(), ROOT_SPAN);
 
     const policy = tracingPolicy().create(mockPolicy, new RequestPolicyOptions());
     await policy.sendRequest(request);