+++ sdk/formrecognizer/ai-form-recognizer/test/internal/tracing.spec.ts
@@ -4,13 +4,12 @@
 import { assert } from "chai";
 import * as sinon from "sinon";
 import { createSpan } from "../../src/tracing";
-import { setTracer } from "@azure/core-tracing";
-import { TestTracer, TestSpan } from "@azure/test-utils";
+import { TestSpan, setTracer } from "@azure/test-utils";
 import { SpanKind, TraceFlags } from "@azure/core-tracing";
 
 describe("tracing.createSpan", () => {
   it("returns a created span with the right metadata", () => {
-    const tracer = new TestTracer();
+    const tracer = setTracer();
     const testSpan = new TestSpan(
       tracer,
       "testing",
@@ -20,7 +19,6 @@ describe("tracing.createSpan", () => {
     const setAttributeSpy = sinon.spy(testSpan, "setAttribute");
     const startSpanStub = sinon.stub(tracer, "startSpan");
     startSpanStub.returns(testSpan);
-    setTracer(tracer);
     const { span } = createSpan("testOperation", {});
     assert.strictEqual(span, testSpan, "Should return mocked span");
     assert.isTrue(startSpanStub.calledOnce);