+++ sdk/storage/storage-blob-changefeed/test/blobchangefeedclient.spec.ts
@@ -6,8 +6,7 @@ import { record, isPlaybackMode, Recorder } from "@azure/test-utils-recorder";
 import { recorderEnvSetup, getBlobChangeFeedClient } from "./utils";
 import { BlobChangeFeedClient, BlobChangeFeedEvent, BlobChangeFeedEventPage } from "../src";
 import { AbortController } from "@azure/abort-controller";
-import { setTracer } from "@azure/core-tracing";
-import { TestTracer } from "@azure/test-utils";
+import { setTracer } from "@azure/test-utils";
 import { Pipeline } from "@azure/storage-blob";
 import { SDK_VERSION } from "../src/utils/constants";
 import { setSpan, context } from "@azure/core-tracing";
@@ -159,8 +158,7 @@ describe("BlobChangeFeedClient", async () => {
   });
 
   it("tracing", async () => {
-    const tracer = new TestTracer();
-    setTracer(tracer);
+    const tracer = setTracer();
     const rootSpan = tracer.startSpan("root");
 
     const pageIter = changeFeedClient.listChanges({