+++ sdk/servicebus/service-bus/test/internal/unit/tracing.spec.ts
@@ -7,7 +7,6 @@ import {
   Context as OTContext,
   getSpanContext,
   getTracer,
-  NoOpSpan,
   setSpanContext,
   SpanOptions,
   SpanStatusCode
@@ -310,7 +309,7 @@ describe("Tracing tests", () => {
     };
 
     it("basic span properties are set", async () => {
-      const fakeParentSpanContext = new NoOpSpan().spanContext();
+      const fakeParentSpanContext = tracer.startSpan("test").spanContext();
 
       createProcessingSpan([], receiverProperties, connectionConfig, {
         tracingOptions: {
@@ -323,7 +322,8 @@ describe("Tracing tests", () => {
       tracer.spanOptions!.kind!.should.equal(SpanKind.CONSUMER);
       getSpanContext(tracer.context!)!.should.equal(fakeParentSpanContext);
 
-      const attributes = tracer.getRootSpans()[0].attributes;
+      const attributes = tracer.getActiveSpans().find((s) => s.name === "Azure.ServiceBus.process")
+        ?.attributes;
 
       attributes!.should.deep.equal({
         "az.namespace": "Microsoft.ServiceBus",