+++ sdk/test-utils/perfstress/src/tests.ts
@@ -49,7 +49,6 @@ export abstract class PerfStressTest<TOptions = {}> {
   public setup?(): void | Promise<void>;
   public cleanup?(): void | Promise<void>;
 
-  public run?(abortSignal?: AbortSignalLike): void;
   public async runAsync?(abortSignal?: AbortSignalLike): Promise<void>;
 
   public getRecordingClient(): RecordingHttpClient {