+++ sdk/monitor/monitor-opentelemetry-exporter/samples/v1/javascript/httpSample.js
@@ -43,9 +43,10 @@ const http = require("http");
  *  HTTP SERVER SETUP
  **********************************************************************/
 /** Starts a HTTP server that receives requests on sample server port. */
+let server;
 function startServer(port) {
   // Creates a server
-  const server = http.createServer(handleRequest);
+  server = http.createServer(handleRequest);
   // Starts the server
   server.listen(port, () => {
     console.log(`Node HTTP listening on ${port}`);
@@ -74,6 +75,8 @@ function handleRequest(request, response) {
     setTimeout(() => {
       span.end();
       response.end("Hello World!");
+      // terminate the process to stop CI pipeline from running forever
+      server.close();
     }, 2000);
   });
 }