+++ sdk/storage/storage-blob/test/containerclient.spec.ts
@@ -3,7 +3,7 @@
 
 import * as assert from "assert";
 import * as dotenv from "dotenv";
-import { TestTracer, SpanGraph } from "@azure/test-utils";
+import { TestTracer, SpanGraph, setTracer } from "@azure/test-utils";
 import {
   bodyToString,
   getBSU,
@@ -20,7 +20,7 @@ import {
   BlobServiceClient
 } from "../src";
 import { Test_CPK_INFO } from "./utils/constants";
-import { context, setSpan, setTracer } from "@azure/core-tracing";
+import { context, setSpan } from "@azure/core-tracing";
 dotenv.config();
 
 describe("ContainerClient", () => {