+++ sdk/core/core-rest-pipeline/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features Added
 
+- `tracingPolicy` will no longer inject invalid traceparent headers if an incorrect tracer implementation is used.
+
 ### Breaking Changes
 
 ### Bugs Fixed